+++ cmd/lsif-go/main.go
@@ -5,7 +5,9 @@ import (
 	"log"
 	"os"
 
+	"github.com/sourcegraph/lsif-go/internal/git"
 	"github.com/sourcegraph/lsif-go/internal/gomod"
+	"github.com/sourcegraph/lsif-go/internal/output"
 )
 
 func init() {
@@ -16,27 +18,49 @@ func init() {
 
 func main() {
 	if err := mainErr(); err != nil {
-		fmt.Fprint(os.Stderr, fmt.Sprintf("\nerror: %v\n", err))
+		fmt.Fprint(os.Stderr, fmt.Sprintf("error: %v\n", err))
 		os.Exit(1)
 	}
 }
 
-func mainErr() error {
+func mainErr() (err error) {
 	if err := parseArgs(os.Args[1:]); err != nil {
 		return err
 	}
 
-	moduleName, err := gomod.ModuleName(moduleRoot, repositoryRemote)
-	if err != nil {
-		return err
+	if !git.Check(moduleRoot) {
+		return fmt.Errorf("module root is not a git repository")
 	}
 
-	dependencies, err := gomod.ListDependencies(moduleRoot, moduleName, moduleVersion)
-	if err != nil {
-		return err
+	defer func() {
+		if err != nil {
+			// Add a new line to all errors except for ones that
+			// come from parsing invalid command line arguments
+			// and basic environment sanity checks.
+			//
+			// We will print progress unconditionally after this
+			// point and we want the error text to be clearly
+			// visible.
+			fmt.Fprintf(os.Stderr, "\n")
+		}
+	}()
+
+	outputOptions := output.Options{
+		Verbosity:      getVerbosity(),
+		ShowAnimations: !noAnimation,
 	}
 
-	return writeIndex(
+	moduleName, err := gomod.ModuleName(moduleRoot, repositoryRemote, outputOptions)
+	if err != nil {
+		return fmt.Errorf("failed to infer module name: %v", err)
+	}
+
+	dependencies, err := gomod.ListDependencies(moduleRoot, moduleName, moduleVersion, outputOptions)
+	if err != nil {
+		return fmt.Errorf("failed to list dependencies: %v", err)
+	}
+
+	if err := writeIndex(
 		repositoryRoot,
 		repositoryRemote,
 		projectRoot,
@@ -44,5 +68,10 @@ func mainErr() error {
 		moduleVersion,
 		dependencies,
 		outFile,
-	)
+		outputOptions,
+	); err != nil {
+		return fmt.Errorf("failed to index: %v", err)
+	}
+
+	return nil
 }